refactor(ui): centralize SiteHeader in @portfolio/ui
Lift the harry/emily header duplicates into a shared SiteHeader. Brand
divergence is captured via opt-in props:

- showLoadingStatus, enableLabDomain, specialPages,
  readingProgressMatchers — harry-only feature flags.
- hideAtPageBottom — emily-only.
- brandVariant ("italic" | "plain"), navLinkClassName for typography.
- staggeredMenu config object passes through to the shared menu.

Both apps reduce to ~50-line wrappers that supply nav items, social
groups, and brand config.
